main.py
#!/usr/bin/env python3
"""
Upwork Job Sniper - Main Application

This script serves as the entry point for the Upwork Job Sniper application.
It initializes the application and starts the job monitoring process.
"""

import asyncio
import logging
import signal
from pathlib import Path

import uvicorn
from dotenv import load_dotenv

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
    handlers=[
        logging.StreamHandler(),
        logging.FileHandler("upwork_sniper.log")
    ]
)
logger = logging.getLogger(__name__)

# Load environment variables
load_dotenv()

class UpworkJobSniper:
    """Main application class for Upwork Job Sniper."""
    
    def __init__(self):
        """Initialize the application."""
        self.should_exit = False
        self.setup_signal_handlers()
    
    def setup_signal_handlers(self):
        """Set up signal handlers for graceful shutdown."""
        signal.signal(signal.SIGINT, self.handle_exit)
        signal.signal(signal.SIGTERM, self.handle_exit)
    
    def handle_exit(self, signum, frame):
        """Handle application shutdown."""
        logger.info("Shutting down gracefully...")
        self.should_exit = True
    
    async def run(self):
        """Run the main application loop."""
        logger.info("Starting Upwork Job Sniper...")
        
        try:
            # Main application loop
            while not self.should_exit:
                # TODO: Implement job fetching and processing
                logger.info("Checking for new jobs...")
                await asyncio.sleep(10)  # Temporary: Replace with actual polling interval
        except Exception as e:
            logger.error(f"An error occurred: {e}", exc_info=True)
        finally:
            logger.info("Upwork Job Sniper has been shut down.")

def main():
    """Initialize and run the application."""
    app = UpworkJobSniper()
    asyncio.run(app.run())

if __name__ == "__main__":
    main()
