requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0
python-dateutil>=2.8.2
pydantic>=2.0.0

# Upwork API
python-upwork>=2.0.0

# OpenAI
openai>=1.0.0

# Notifications
python-pushover>=1.1.0

# Web Interface
gradio>=3.50.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pytest>=7.3.1
pytest-cov>=4.0.0
pre-commit>=3.3.0

# Async
httpx>=0.24.0
asyncio>=3.4.3

# Data
orjson>=3.9.0
pandas>=2.0.0
